moved the plot to visualisation
